+++ apps/explorer/src/explorer/components/OrderWidget/index.tsx
@@ -6,10 +6,21 @@ import { useOrderTrades } from '../../../hooks/useOperatorTrades'
 import { useSanitizeOrderIdAndUpdateUrl } from '../../../hooks/useSanitizeOrderIdAndUpdateUrl'
 import { RedirectToNetwork, useNetworkId } from '../../../state/network'
 import { ORDER_QUERY_INTERVAL } from '../../const'
+import { useTable } from '../TokensTableWidget/useTable'
+
+const RESULTS_PER_PAGE = 10
 
 export const OrderWidget: React.FC = () => {
   const networkId = useNetworkId()
   const orderId = useSanitizeOrderIdAndUpdateUrl()
+
+  const {
+    state: tableState,
+    setPageSize,
+    setPageOffset,
+    handleNextPage,
+    handlePreviousPage,
+  } = useTable({ initialState: { pageOffset: 0, pageSize: RESULTS_PER_PAGE } })
 
   const {
     order,
@@ -17,7 +28,15 @@ export const OrderWidget: React.FC = () => {
     errors,
     errorOrderPresentInNetworkId,
   } = useOrderAndErc20s(orderId, ORDER_QUERY_INTERVAL)
-  const { trades, error, isLoading: areTradesLoading } = useOrderTrades(order)
+  const {
+    trades,
+    error,
+    isLoading: areTradesLoading,
+    hasNextPage,
+  } = useOrderTrades(order, tableState.pageOffset, tableState.pageSize)
+
+  // eslint-disable-next-line react-hooks/immutability
+  tableState['hasNextPage'] = hasNextPage
 
   if (error) {
     // eslint-disable-next-line react-hooks/immutability
@@ -35,6 +54,11 @@ export const OrderWidget: React.FC = () => {
       isOrderLoading={isOrderLoading}
       areTradesLoading={areTradesLoading}
       errors={errors}
+      tableState={tableState}
+      setPageSize={setPageSize}
+      setPageOffset={setPageOffset}
+      handleNextPage={handleNextPage}
+      handlePreviousPage={handlePreviousPage}
     />
   )
 }